add an extra check to prevent deleting directories with documentFile

+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/MainActivity.kt
@@ -202,7 +202,7 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
                 val document = getFileDocument(file.absolutePath, mConfig.treeUri)
 
                 // double check we have the uri to the proper file path, not some parent folder
-                if (document.uri.toString().endsWith(file.absolutePath.getFilenameFromPath()))
+                if (document.uri.toString().endsWith(file.absolutePath.getFilenameFromPath()) && !document.isDirectory)
                     document.delete()
             }
         } else {
